Refactor FacebookCommerce extension to controller-based structure
